remove special context per feature struct
instead update the macro to generate the suites directly
