+++ features/config.exs
@@ -1,8 +1,7 @@
 defmodule WhiteBread.Example.Config do
   use WhiteBread.SuiteConfiguration
 
-  context_per_feature on: true,
-                      namespace_prefix: WhiteBread.TestContextPerFeature,
+  context_per_feature namespace_prefix: WhiteBread.TestContextPerFeature,
                       entry_feature_path: "features/test_context_per_feature/"
 
   suite name:          "Default context",
