--- lib/white_bread/context_per_feature.ex
@@ -1,5 +1,0 @@
-defmodule WhiteBread.ContextPerFeature do
-  defstruct on: false,
-            namespace_prefix: WhiteBread.Context,
-            entry_feature_path: "features/"
-end
