+++ test/suite/context_per_feature_test.exs
@@ -1,17 +1,15 @@
-defmodule WhiteBread.CommandLine.ContextPerFeatureTest do
+defmodule WhiteBread.Suite.ContextPerFeatureTest do
   use ExUnit.Case
-  alias WhiteBread.CommandLine
-  alias WhiteBread.ContextPerFeature
   alias WhiteBread.Suite
 
+  alias WhiteBread.Suite.ContextPerFeature
+
   test "Build a list of suites" do
-    {status, actual} = CommandLine.ContextPerFeature.build_suites(%ContextPerFeature{
-        on: true,
+    actual = ContextPerFeature.build_suites(%{
         namespace_prefix: WhiteBread,
         entry_feature_path: "features/"
       })
 
-    assert status == :ok
     assert Enum.member?(actual, %Suite{
             name: "Example1",
             context: WhiteBread.Example1Context,
@@ -19,23 +17,12 @@ defmodule WhiteBread.CommandLine.ContextPerFeatureTest do
           })
   end
 
-  test "Context per feature is turned off" do
-    {status, _message} = CommandLine.ContextPerFeature.build_suites(%ContextPerFeature{
-        on: false,
-        namespace_prefix: WhiteBread,
-        entry_feature_path: "features/"
-      })
-
-    assert status == :error
-  end
-
   test "Build a suite struct" do
-    actual = CommandLine.ContextPerFeature.build_suite(%ContextPerFeature{ 
-        on: true,
+    actual = ContextPerFeature.build_suite(%{
         namespace_prefix: WhiteBread,
         entry_feature_path: "features/"
       }, "features/something_special.feature")
-    
+
     assert actual == %Suite{
             name: "SomethingSpecial",
             context: WhiteBread.SomethingSpecialContext,
